Conflicts:
	Assets/Scenes/TestScene_PlayerPrefab.unity

+++ Assets/Script/AttackInfo.cs
@@ -3,29 +3,29 @@ using System.Collections;
 using Cradle;
 
 namespace Cradle{
-public class AttackInfo : MonoBehaviour, IInfoController{
-	private Transform attacker;
-	public AttackInfoController aIController;
-		
-	public void OnEnable() {
-		aIController.SetInfoController (this);
-	}
+	public class AttackInfo : MonoBehaviour, IInfoController{
+		private Transform attacker;
+		public AttackInfoController aIController;
+			
+		public void OnEnable() {
+			aIController.SetInfoController (this);
+		}
 
-	public int GetAttackPower(){
-		return aIController.getAttackPower();
-	}
+		public float GetAttackPower(){
+			return aIController.getAttackPower();
+		}
 
-	public int SetAttackPower(int atk){
-			return aIController.setAttackPower (atk);
-	}
+		public float SetAttackPower(float atk){
+				return aIController.setAttackPower (atk);
+		}
 
-	public int SetAttackBoostPower(int atk){
-			return aIController.setAttackBoostPower (atk);
-	}
+		public float SetAttackBoostPower(float atk){
+				return aIController.setAttackBoostPower (atk);
+		}
 
-	public Transform SetAttacker(Transform atk){
-		return this.attacker = atk;
-	}
+		public Transform SetAttacker(Transform atk){
+			return this.attacker = atk;
+		}
 
- }
+	 }
 }
